pyproject.toml
[project]
name = "edmi-app"
version = "0.1.0"
description = "Aplicación para Estudios de Mercado Inmobiliario"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "groq>=0.4.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "PyMuPDF>=1.23.0",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["etl", "services"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
